refactor(front): relocate PageNav component
