+++ front/src/lib/components/general/PageNav/PageNav.svelte
@@ -1,6 +1,6 @@
 <script>
   import TableOfContents from '@components/posts/TableOfContents.svelte';
-  import ClosePageNav from '@components/icons/ClosePageNav.svelte';
+  import ClosePageNav from '@components/general/PageNav/ClosePageNav.svelte';
   import { categoryClick, relatedClick } from '@utils/uiHelpers.js';
   import { focusTrap } from '@utils/actions.js';
   import ListIcon from '~icons/gg/list';
